+++ CHANGELOG.rst
@@ -6,10 +6,21 @@ The following is a summary of changes and improvements to
 any necessary information about installation or upgrade notes.
 
 
+1.3
+---
+
+* Tutorial updated to be compatible with Django 1.8 thanks to
+  `jaska @chfw <https://github.com/chfw>`_.
+* New class-based view :class:`eulfedora.views.RawDatastreamView`,
+  equivalent to :meth:`eulfedora.views.raw_datastream`.
+* Access to historical versions of datastreams now available in
+  :meth:`eulfedora.models.DigitalObject.getDatastreamObject` and
+  :meth:`eulfedora.views.raw_datastream`.
+
 1.2
 ---
 
-* Change checksum handling to cue Fedora to auto-generating checksums
+* Change checksum handling to cue Fedora to auto-generate checksums
   on ingest.
 * **Recommended**: Fedora 3.7+ for automatic checksum support on ingest